Hopefully fixed cobot from running into self + design changes
