+++ src/main/java/org/example/Blackboard.java
@@ -4,22 +4,22 @@ import java.util.concurrent.ConcurrentLinkedQueue;
 
 public class Blackboard {
     private static Blackboard instance;
-    private final ConcurrentLinkedQueue<ContourParser.Point> points = new ConcurrentLinkedQueue<>();
+    private final ConcurrentLinkedQueue<Point> points = new ConcurrentLinkedQueue<>();
 
     private Blackboard() {}
 
-    public static synchronized Blackboard getInstance() {
+    public static Blackboard getInstance() {
         if (instance == null) {
             instance = new Blackboard();
         }
         return instance;
     }
 
-    public void addPoint(ContourParser.Point point) {
+    public void addPoint(Point point) {
         points.add(point);
     }
 
-    public ContourParser.Point getPoint() {
+    public Point getPoint() {
         return points.poll();
     }
