+++ src/main/java/org/example/Point.java
@@ -1,0 +1,19 @@
+package org.example;
+
+public class Point {
+    private final double X;
+    private final double Y;
+
+    public Point(double x, double y) {
+        X = x;
+        Y = y;
+    }
+
+    public double getX() {
+        return X;
+    }
+
+    public double getY() {
+        return Y;
+    }
+}
